Run ssh key pair gen command once
The documentation already notes that this is not
a per-container setting, so instead of fighting
to implement a reliable per-container setting
that nobody may use, go with the initial idea
of sharing the key pair for access to all
containers. If the other approach is needed
later I can implement the feature then.

refs #43